add some changes

+++ src/Clinic.Domain/Entities/Doctor.cs
@@ -11,7 +11,7 @@ namespace Clinic.Domain.Entities
         public Guid Id { get; set; }
         public string Name { get; set; }
         public string Address { get; set; }
-        public long StartWork { get; set; }
+        public int StartWork { get; set; }
         public string TUsername { get; set; }
         public string? PicturePath { get; set; }
